requirements.txt
Authlib==0.15.5
bcrypt==3.2.0
cachelib==0.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
dnspython==2.1.0
Flask==2.0.2
Flask-PyMongo==2.3.0
Flask-Session==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauth==1.0.1
oauthlib==3.1.1
pathlib==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
